Executable added
Added the executable file showing the pool game at work (please excuse the watermark - all the features I needed for this were in the free version of Game Maker!).

Controls for the white ball: Left/right arrows steer the aim, and holding the up arrow charges the shot; release at any time to fire with that much force. The N/M keys can be used for finer aiming.